requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
bcrypt==4.0.1

# Caching & Sessions
Flask-Caching==2.1.0
Flask-Session==0.5.0
redis==5.0.1

# Mail
Flask-Mail==0.9.1

# AI Integration
openai==1.3.5
langchain==0.0.340
python-dotenv==1.0.0

# File handling
Pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# Forms and validation
email-validator==2.1.0

# Language support
Babel==2.13.1
Flask-Babel==4.0.0

# Development tools
Flask-DebugToolbar==0.13.1

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
bleach==6.1.0